requirements.txt
aniso8601==9.0.1
astroid==3.2.4
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
dill==0.3.8
Flask==3.0.3
flask-swagger==0.2.14
idna==3.8
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
json-queries==0.1.16
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mccabe==0.7.0
multipledispatch==1.0.0
numpy==2.1.0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pyperclip==1.9.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
soupsieve==2.6
tomlkit==0.13.2
urllib3==2.2.2
Werkzeug==3.0.4
